Minor fixes in indexer
