+++ seismicpro/src/indexer.py
@@ -3,14 +3,6 @@ from itertools import chain
 import numpy as np
 
 
-def unique_indices_sorted(arr):
-    """Return indices of the first occurrences of the unique values in a sorted array."""
-    mask = np.empty(len(arr), dtype=np.bool_)
-    np.any(arr[1:] != arr[:-1], axis=1, out=mask[1:])
-    mask[0] = True
-    return np.where(mask)[0]
-
-
 class BaseIndexer:
     def __init__(self, index):
         self.index = index
@@ -25,6 +17,7 @@ class TraceIndexer(BaseIndexer):
     def __init__(self, index):
         super().__init__(index)
         self.unique_indices = index
+        _ = self.get_loc(index[:1])  # Warmup call to `get_loc`: the first call is way slower than the following ones
 
     def get_loc(self, index):
         return self.index.get_indexer(index)
@@ -33,10 +26,9 @@ class TraceIndexer(BaseIndexer):
 class GatherIndexer(BaseIndexer):
     def __init__(self, index):
         super().__init__(index)
-        unique_indices_pos = unique_indices_sorted(index.to_frame().values)
+        unique_indices_pos = np.where(~index.duplicated())[0]
         ix_start = unique_indices_pos
         ix_end = chain(unique_indices_pos[1:], [len(index)])
-
         self.unique_indices = index[unique_indices_pos]
         self.index_to_headers_pos = {ix: range(*args) for ix, *args in zip(self.unique_indices, ix_start, ix_end)}
